ProductOffering circle. Added debet and credit cards support. Some changes in card controller

+++ src/main/java/com/presentation/demo/controller/BillController.java
@@ -1,7 +1,6 @@
 package com.presentation.demo.controller;
 
 import com.presentation.demo.model.Bill;
-import com.presentation.demo.model.Card;
 import com.presentation.demo.model.DateBalanceHistory;
 import com.presentation.demo.model.User;
 import com.presentation.demo.service.bill.BillService;
@@ -16,7 +15,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.*;
 
 @Controller
@@ -36,7 +35,7 @@ public class BillController {
 
     @GetMapping("/createbill/{id}")
     @ResponseBody
-    public String createBill(@PathVariable("id") Integer id) {
+    public String createBill(@PathVariable("id") Long id) {
 
         Random rand = new Random();
         Bill bill = new Bill();
@@ -63,7 +62,7 @@ public class BillController {
         Bill bill = billService.findBillById(billId);
         model.addAttribute("billId",bill.getId());
         model.addAttribute("holderId",bill.getHolder().getId());
-        Map<Date, BigInteger> dateBigIntegerMap = new TreeMap<>();
+        Map<Date, BigDecimal> dateBigIntegerMap = new TreeMap<>();
         List<DateBalanceHistory> dateBalanceHistories = dateBalanceHistoryService.findDateBalanceHistoriesByBill(bill);
         for (DateBalanceHistory dateBalanceHistory: dateBalanceHistories) {
             dateBigIntegerMap.put(dateBalanceHistory.getDate(), dateBalanceHistory.getBalance());
